applications/crypto-ingress/pyproject.toml
[tool.poetry]
name = "crypto-ingress"
version = "0.1.0"
description = ""
authors = ["Max Maton <me@mmaton.com>"]
readme = "README.md"

[virtualenvs]
in-project = true

[tool.poetry.dependencies]
python = ">=3.11,<=3.14"
influxdb-client = "^1.40.0"
python-kraken-sdk = "^2.1.1"
environs = "^10.3.0"
paho-mqtt = "^2.0.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
